Coding HW 9- Due 10/16
Attached are png files for the radioactive devay graph (the ipynb file also gives a description of what I did for the decay graph), as well as the final numerical results for the Let's Make A Deal problem. _stay is when the computer styayed with the initial guess, _switch is when the computer switched its final answer. The ipynb file also described the results of the game and what is statistically the best thing to do when playing.